FileFormatBase: Split FileFormat class
